Alteration to definition of done
While "ready to deploy" is better than many other definitions of done, the work is only truly finished when it has been deployed to production successfully and the change verified as working correctly there. Allowing teams to call "done" earlier in the software delivery process creates a backlog/buffer of unfinished work a.k.a. "waste" in lean terminology.