Clarify roles and responsibilities
Not everyone is (or should be) responsible for or empowered to change the practices and processes in place in order to move us towards our desired future state. In particular, the framework is lacking a clear explanation of who is responsible for what and how they can be held accountable for upholding the principles. Therefore I propose a new section setting out what is expected from roles such as Developer, Delivery Manager, Product Owner, and so on. For example, if we consider Developers to be ultimately responsible for quality then we must say this, and we must also also explicitly state that this responsibility brings with it the powers required to fulfil the responsibility e.g. control over workload, pushback in the face of "deadlines", permission to refactor if required and so on.