Add nuance to outsource-bottom-up
Previously the outsource-bottom-up.md guidance was tilted strongly towards
promoting cloud functions at the expense of any other compute model, whereas
the CCoE policy is more relaxed.  The previous wording ignores difficulties
inherent in the cloud function model which are worth expanding on so that the
reasoning behind the recommendations is clearer.